Honor DRACUT_CPIO_NO_MTIME environment variable
The OSTree system uses content-addressed storage, and when initramfs
images generated by dracut are in the tree, because the cpio format
contains modification/ctime, it causes a different checksum every time
it's created.

Allow builders to avoid that by using a stable reference time.  This
is not the default, although maybe it should be?

I didn't add a command-line argument because it'd break on older
versions of dracut.