bridging: generate ifcfg files for both the bridge and interfaces
When a bridged interface is configured to start up on boot via the
kernel command line options, the resulting ifcfg files produced by
the 45ifcfg dracut module are incorrect.  Atlthough the interface is
brought up correctly on start up, and functions as expected, when the
configuration file is imported into the booted system it prevents the
bridged interface from being successfully restarted.

This patch rewrites the section of 45ifcfg responsible for the creation
of the required ifcfg files.

Bug #1123552 https://bugzilla.redhat.com/show_bug.cgi?id=1123552