dmsquash-live-root: Set default transient overlay size to 32768 MiB.
The most common complaint about LiveOS systems are crashes due to
snapshot overlays that overflow or get invalidated by excessive data
storage.

The new PO, Persistent with Overflow, DM target setup parameter
avoids the severe impact of an invalidated snapshot, but overflows
would become less frequent if the default overlay_size was larger
than 512 MiB--a value probably set at a time when 1024 MiB of RAM
was a common memory size in machines.

The transient overlay is a sparse file in a TMPFS, so it only
consumes space as more space is needed.

32 GiB is larger than common TMPFSs, so the risk of having the root
filesystem silently set to read-only status due to overlay overflow
is traded for the risk that one will receive overt warnings of low
memory when the overlay storage approaches the available memory
resources.