use more portable /usr/bin/env in shebangs
This commit changes all shebangs to use `#!/usr/bin/env bash` instead of
expecting bash to be located in `/bin`. This is especially important
on Nixos, which trades LFS for a more safer pure package manager. Also
the trend in some distributions is to give up `/bin/` for putting all
binaries into `/usr/bin/`.