change the default value of rd.timeout from 0 to 300
Currently rd.timeout default value is 0, this will result in the
systemd job waiting for the target forever, and cause the boot
process hang forever, it's better to wait some default time other
than forever.

We met an issue under kdump, the dump target broke for some reason.
Because wait_for_dev() is invoked due to "--hostonly-cmdline" is
specified when making kdump initramfs in the 1st kernel, we have to
add "rd.timeout=X" to the 1st kernel using grub manually, which is
unacceptable for customers.

This patch sets the default value of rd.timeout to 300s, which is
longer than the default value of rd.retry(180s).

Signed-off-by: Xunlei Pang <xlpang@redhat.com>